+++ packages/at_secondary_server/test/enroll_data_store_value_test.dart
@@ -6,13 +6,7 @@ import 'package:at_persistence_secondary_server/at_persistence_secondary_server.
 import 'package:at_secondary/src/connection/inbound/inbound_connection_metadata.dart';
 import 'package:at_secondary/src/constants/enroll_constants.dart';
 import 'package:at_secondary/src/enroll/enroll_datastore_value.dart';
-import 'package:at_secondary/src/utils/handler_util.dart';
-import 'package:at_secondary/src/verb/handler/delete_verb_handler.dart';
-import 'package:at_secondary/src/verb/handler/enroll_verb_handler.dart';
-import 'package:at_secondary/src/verb/handler/totp_verb_handler.dart';
-import 'package:at_secondary/src/verb/handler/update_verb_handler.dart';
 import 'package:test/test.dart';
-
 import 'test_utils.dart';
 
 void main() {
@@ -31,30 +25,12 @@ void main() {
       expect(enrollApprovalJson['state'], 'requested');
     });
 
-    test('enroll namespace object fromJson', () {
-      final enrollNamespaceJson = {'name': 'buzz', 'access': 'r'};
-      final enrollNamespaceObject =
-          EnrollNamespace.fromJson(enrollNamespaceJson);
-      expect(enrollNamespaceObject, isA<EnrollNamespace>());
-      expect(enrollNamespaceObject.name, 'buzz');
-      expect(enrollNamespaceObject.access, 'r');
-    });
-
-    test('enroll namespace object toJson', () {
-      final enrollNamespace = EnrollNamespace('wavi', 'rw');
-      final enrollNamespaceJson = enrollNamespace.toJson();
-      expect(enrollNamespaceJson['name'], 'wavi');
-      expect(enrollNamespaceJson['access'], 'rw');
-    });
-
     test('enroll data store value object toJson', () {
-      final enrollNamespace1 = EnrollNamespace('wavi', 'rw');
-      final enrollNamespace2 = EnrollNamespace('buzz', 'r');
+      var namespaceMap = {'wavi': 'rw', 'buzz': 'r'};
       final enrollApproval = EnrollApproval('requested');
-      final namespaceList = [enrollNamespace1, enrollNamespace2];
       final enrollDataStoreValue =
           EnrollDataStoreValue('123', 'testclient', 'iphone', 'mykey')
-            ..namespaces = namespaceList
+            ..namespaces = namespaceMap
             ..approval = enrollApproval
             ..requestType = EnrollRequestType.newEnrollment;
       final enrollJson = enrollDataStoreValue.toJson();
@@ -63,8 +39,8 @@ void main() {
       expect(enrollJson['deviceName'], 'iphone');
       expect(enrollJson['apkamPublicKey'], 'mykey');
       expect(enrollJson['requestType'], 'newEnrollment');
-      expect(enrollJson['namespaces'][0], enrollNamespace1);
-      expect(enrollJson['namespaces'][1], enrollNamespace2);
+      expect(enrollJson['namespaces']['wavi'], 'rw');
+      expect(enrollJson['namespaces']['buzz'], 'r');
       expect(enrollJson['approval'], enrollApproval);
     });
     test('enroll data store value object fromJson', () {
@@ -72,10 +48,7 @@ void main() {
         'sessionId': '123',
         'appName': 'testclient',
         'deviceName': 'iphone',
-        'namespaces': [
-          {'name': 'wavi', 'access': 'rw'},
-          {'name': 'buzz', 'access': 'r'}
-        ],
+        'namespaces': {'wavi': 'rw', 'buzz': 'r'},
         'apkamPublicKey': 'mykey',
         'requestType': 'newEnrollment',
         'approval': {'state': 'requested'}
@@ -83,7 +56,7 @@ void main() {
       final enrollValueObject = EnrollDataStoreValue.fromJson(enrollJson);
       expect(enrollValueObject, isA<EnrollDataStoreValue>());
       expect(enrollValueObject.approval, isA<EnrollApproval>());
-      expect(enrollValueObject.namespaces, isA<List<EnrollNamespace>>());
+      expect(enrollValueObject.namespaces, isA<Map<String, String>>());
       expect(enrollValueObject.sessionId, '123');
       expect(enrollValueObject.appName, 'testclient');
       expect(enrollValueObject.deviceName, 'iphone');