+++ packages/at_secondary_server/test/scan_verb_test.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:at_commons/at_commons.dart';
 import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
 import 'package:at_secondary/src/caching/cache_manager.dart';
 import 'package:at_secondary/src/connection/inbound/dummy_inbound_connection.dart';
@@ -8,6 +9,7 @@ import 'package:at_secondary/src/connection/inbound/inbound_connection_metadata.
 import 'package:at_secondary/src/connection/outbound/outbound_client_manager.dart';
 import 'package:at_secondary/src/notification/notification_manager_impl.dart';
 import 'package:at_secondary/src/notification/stats_notification_service.dart';
+import 'package:at_secondary/src/utils/handler_util.dart';
 import 'package:at_secondary/src/utils/secondary_util.dart';
 import 'package:at_secondary/src/verb/executor/default_verb_executor.dart';
 import 'package:at_secondary/src/verb/handler/response/default_response_handler.dart';
@@ -17,10 +19,11 @@ import 'package:at_secondary/src/verb/manager/response_handler_manager.dart';
 import 'package:at_secondary/src/verb/manager/verb_handler_manager.dart';
 import 'package:at_server_spec/at_server_spec.dart';
 import 'package:at_server_spec/at_verb_spec.dart';
-import 'package:test/test.dart';
-import 'package:at_secondary/src/utils/handler_util.dart';
-import 'package:at_commons/at_commons.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:test/test.dart';
+import 'package:uuid/uuid.dart';
+
+import 'test_utils.dart';
 
 // Global variable to assert the scan responses from the mock response handlers
 String scanResponse = '';
@@ -180,5 +183,236 @@ void main() {
       expect(scanResponseList.contains('public:__phone.wavi@alice'), false);
       expect(scanResponseList.contains('_mobile.wavi@alice'), false);
     });
+  });
+
+  group('A group of APKAM enrollment tests', () {
+    late ScanVerbHandler scanVerbHandler;
+    late ResponseHandlerManager mockResponseHandlerManager;
+    setUp(() async {
+      await verbTestsSetUp();
+    });
+
+    test(
+        'A test to verify scan does not return the enrollment keys when enrollment namespace has __manage',
+        () async {
+      inboundConnection.getMetaData().isAuthenticated = true;
+      inboundConnection.getMetaData().sessionID = 'dummy_session';
+      var enrollmentId = Uuid().v4();
+      final enrollJson = {
+        'sessionId': '123',
+        'appName': 'wavi',
+        'deviceName': 'pixel',
+        'namespaces': {'__manage': 'r', 'wavi': 'r'},
+        'apkamPublicKey': 'testPublicKeyValue',
+        'requestType': 'newEnrollment',
+        'approval': {'state': 'approved'}
+      };
+      var keyName = '$enrollmentId.new.enrollments.__manage@alice';
+      await secondaryKeyStore.put(
+          keyName, AtData()..data = jsonEncode(enrollJson));
+      await secondaryKeyStore.put(
+          'public:firstName.wavi$alice', AtData()..data = 'alice');
+
+      scanVerbHandler = ScanVerbHandler(
+          secondaryKeyStore, mockOutboundClientManager, cacheManager);
+      mockResponseHandlerManager = MockResponseHandlerManager();
+      // Set enrollmentId to the inboundConnection to mimic the APKAM auth
+      (inboundConnection.getMetaData() as InboundConnectionMetadata)
+          .enrollmentId = enrollmentId;
+      scanVerbHandler.responseManager = mockResponseHandlerManager;
+      await scanVerbHandler.process('scan', inboundConnection);
+      List scanResponseList = jsonDecode(scanResponse);
+      expect(scanResponseList, isNotEmpty);
+      expect(
+          scanResponseList
+              .contains('$enrollmentId.new.enrollments.__manage@alice'),
+          false);
+    });
+
+    test(
+        'A test to verify scan returns only the keys whose namespaces are authorized in enrollment request',
+        () async {
+      inboundConnection.getMetaData().isAuthenticated = true;
+      inboundConnection.getMetaData().sessionID = 'dummy_session';
+      var enrollmentId = Uuid().v4();
+      final enrollJson = {
+        'sessionId': '123',
+        'appName': 'wavi',
+        'deviceName': 'pixel',
+        'namespaces': {'__manage': 'r', 'wavi': 'r'},
+        'apkamPublicKey': 'testPublicKeyValue',
+        'requestType': 'newEnrollment',
+        'approval': {'state': 'approved'}
+      };
+      var keyName = '$enrollmentId.new.enrollments.__manage@alice';
+      await secondaryKeyStore.put(
+          keyName, AtData()..data = jsonEncode(enrollJson));
+      // Insert key with wavi and buzz namespace
+      await secondaryKeyStore.put(
+          'firstName.wavi$alice', AtData()..data = 'alice');
+      await secondaryKeyStore.put(
+          'mobileNumber.buzz$alice', AtData()..data = '+1 434 543 3232');
+
+      scanVerbHandler = ScanVerbHandler(
+          secondaryKeyStore, mockOutboundClientManager, cacheManager);
+      mockResponseHandlerManager = MockResponseHandlerManager();
+      // Set enrollmentId to the inboundConnection to mimic the APKAM auth
+      (inboundConnection.getMetaData() as InboundConnectionMetadata)
+          .enrollmentId = enrollmentId;
+      scanVerbHandler.responseManager = mockResponseHandlerManager;
+      await scanVerbHandler.process('scan', inboundConnection);
+      List scanResponseList = jsonDecode(scanResponse);
+      expect(scanResponseList.length, 1);
+      expect(scanResponseList[0], 'firstname.wavi$alice');
+    });
+
+    test(
+        'A test to verify scan returns enrollment keys on a CRAM authenticated connection',
+        () async {
+      inboundConnection.getMetaData().isAuthenticated = true;
+      inboundConnection.getMetaData().sessionID = 'dummy_session';
+      inboundConnection.getMetaData().authType = AuthType.cram;
+      var enrollmentId = Uuid().v4();
+      final enrollJson = {
+        'sessionId': '123',
+        'appName': 'wavi',
+        'deviceName': 'pixel',
+        'namespaces': {'__manage': 'r', 'wavi': 'r'},
+        'apkamPublicKey': 'testPublicKeyValue',
+        'requestType': 'newEnrollment',
+        'approval': {'state': 'approved'}
+      };
+      var keyName = '$enrollmentId.new.enrollments.__manage@alice';
+      await secondaryKeyStore.put(
+          keyName, AtData()..data = jsonEncode(enrollJson));
+
+      scanVerbHandler = ScanVerbHandler(
+          secondaryKeyStore, mockOutboundClientManager, cacheManager);
+      mockResponseHandlerManager = MockResponseHandlerManager();
+      scanVerbHandler.responseManager = mockResponseHandlerManager;
+      await scanVerbHandler.process('scan', inboundConnection);
+      List scanResponseList = jsonDecode(scanResponse);
+      expect(scanResponseList.length, 1);
+      expect(
+          scanResponseList[0], '$enrollmentId.new.enrollments.__manage$alice');
+    });
+
+    test('A test to verify enrollment has *:rw access', () async {
+      var enrollmentId = Uuid().v4();
+      inboundConnection.getMetaData().isAuthenticated = true;
+      inboundConnection.getMetaData().sessionID = 'dummy_session';
+      (inboundConnection.getMetaData() as InboundConnectionMetadata)
+          .enrollmentId = enrollmentId;
+
+      final enrollJson = {
+        'sessionId': '123',
+        'appName': 'wavi',
+        'deviceName': 'pixel',
+        'namespaces': {'*': 'rw'},
+        'apkamPublicKey': 'testPublicKeyValue',
+        'requestType': 'newEnrollment',
+        'approval': {'state': 'approved'}
+      };
+      var keyName = '$enrollmentId.new.enrollments.__manage@alice';
+      await secondaryKeyStore.put(
+          keyName, AtData()..data = jsonEncode(enrollJson));
+
+      scanVerbHandler = ScanVerbHandler(
+          secondaryKeyStore, mockOutboundClientManager, cacheManager);
+      mockResponseHandlerManager = MockResponseHandlerManager();
+      scanVerbHandler.responseManager = mockResponseHandlerManager;
+      await scanVerbHandler.process('scan', inboundConnection);
+      List scanResponseList = jsonDecode(scanResponse);
+      expect(scanResponseList[0].toString().startsWith(enrollmentId), true);
+    });
+
+    test(
+        'A test to verify scan returns all keys when enrollment has *:rw access',
+        () async {
+      var enrollmentId = Uuid().v4();
+      inboundConnection.getMetaData().isAuthenticated = true;
+      inboundConnection.getMetaData().sessionID = 'dummy_session';
+      (inboundConnection.getMetaData() as InboundConnectionMetadata)
+          .enrollmentId = enrollmentId;
+
+      final enrollJson = {
+        'sessionId': '123',
+        'appName': 'wavi',
+        'deviceName': 'pixel',
+        'namespaces': {'*': 'rw'},
+        'apkamPublicKey': 'testPublicKeyValue',
+        'requestType': 'newEnrollment',
+        'approval': {'state': 'approved'}
+      };
+      var keyName = '$enrollmentId.new.enrollments.__manage@alice';
+      await secondaryKeyStore.put(
+          keyName, AtData()..data = jsonEncode(enrollJson));
+
+      await secondaryKeyStore.put(
+          'public:phone.wavi@alice', AtData()..data = '+455 675 6765');
+      await secondaryKeyStore.put(
+          '@bob:firstName.atmosphere@alice', AtData()..data = 'Alice');
+      await secondaryKeyStore.put(
+          'mobile.buzz@alice', AtData()..data = '+878 787 7679');
+
+      scanVerbHandler = ScanVerbHandler(
+          secondaryKeyStore, mockOutboundClientManager, cacheManager);
+      mockResponseHandlerManager = MockResponseHandlerManager();
+      scanVerbHandler.responseManager = mockResponseHandlerManager;
+      await scanVerbHandler.process('scan', inboundConnection);
+      List scanResponseList = jsonDecode(scanResponse);
+      expect(
+          scanResponseList
+              .contains('$enrollmentId.new.enrollments.__manage@alice'),
+          true);
+      expect(
+          scanResponseList.contains('@bob:firstname.atmosphere@alice'), true);
+
+      expect(scanResponseList.contains('mobile.buzz@alice'), true);
+      expect(scanResponseList.contains('public:phone.wavi@alice'), true);
+    });
+
+    test(
+        'A test to verify multiple app access in enrollment buzz:r, wavi:rw, atmosphere:rw',
+        () async {
+      var enrollmentId = Uuid().v4();
+      inboundConnection.getMetaData().isAuthenticated = true;
+      inboundConnection.getMetaData().sessionID = 'dummy_session';
+      (inboundConnection.getMetaData() as InboundConnectionMetadata)
+          .enrollmentId = enrollmentId;
+
+      final enrollJson = {
+        'sessionId': '123',
+        'appName': 'wavi',
+        'deviceName': 'pixel',
+        'namespaces': {'buzz': 'r', 'wavi': 'rw', 'atmosphere': 'rw'},
+        'apkamPublicKey': 'testPublicKeyValue',
+        'requestType': 'newEnrollment',
+        'approval': {'state': 'approved'}
+      };
+      var keyName = '$enrollmentId.new.enrollments.__manage@alice';
+      await secondaryKeyStore.put(
+          keyName, AtData()..data = jsonEncode(enrollJson));
+      // Inserting wavi
+      await secondaryKeyStore.put(
+          'phone.wavi$alice', AtData()..data = '+455 677 8789');
+      // Inserting buzz
+      await secondaryKeyStore.put(
+          'mobile.buzz$alice', AtData()..data = '+544 545 4545');
+      // Inserting atmosphere
+      await secondaryKeyStore.put(
+          'firstname.atmosphere$alice', AtData()..data = 'alice');
+
+      scanVerbHandler = ScanVerbHandler(
+          secondaryKeyStore, mockOutboundClientManager, cacheManager);
+      mockResponseHandlerManager = MockResponseHandlerManager();
+      scanVerbHandler.responseManager = mockResponseHandlerManager;
+      await scanVerbHandler.process('scan', inboundConnection);
+      List scanResponseList = jsonDecode(scanResponse);
+      expect(scanResponseList[0], 'firstname.atmosphere$alice');
+      expect(scanResponseList[1], 'mobile.buzz$alice');
+      expect(scanResponseList[2], 'phone.wavi$alice');
+    });
+    tearDown(() async => await verbTestsTearDown());
   });
 }