+++ packages/at_secondary_server/CHANGELOG.md
@@ -1,7 +1,16 @@
+## 3.0.35
+- chore: Upgraded at_persistence_secondary_server to 3.0.57 for memory optimization in commit log
+- feat: APKAM keys verb implementation
+- feat: Implementation changes for APKAM latest specification
+- Allow lookup verb for only authorized namespaces when authenticated via APKAM
+## 3.0.34
+- chore: Upgraded at_persistence_spec to 2.0.14
+- chore: Upgraded at_persistence_secondary_server to 3.0.56
 ## 3.0.33
-- feat: modified monitor verb handler to process self notification for apkam
-- chore: upgraded at_persistence_secondary_server to 3.0.54, at_server_spec to 3.0.13, at_commons to 3.0.50 and at_utils 3.0.14
-- feat: apkam enroll verb handler implementation
+- feat: Modified monitor verb handler to process self notification for APKAM
+- chore: Upgraded at_persistence_secondary_server to 3.0.55 for memory optimization
+- chore: Upgraded at_server_spec to 3.0.13, at_commons to 3.0.50 and at_utils 3.0.14
+- feat: APKAM enroll verb handler implementation
 ## 3.0.32
 - fix: Enhance stats verb to return latest commitEntry of each key
 - chore: Ignore melos files